% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kmers.R
\name{get_kmers}
\alias{get_kmers}
\title{Generate kmers from long peptide sequence}
\usage{
get_kmers(sequence, kmer_length = 12, overlap = 6)
}
\arguments{
\item{sequence}{sequence of a single peptide}

\item{kmer_length}{length of desired kmer}

\item{overlap}{number of overlapped residuals between two consecutive kmers}
}
\value{
a character vector that contains all kmers generated from the input peptide sequence
}
\description{
given a sequence, return all kmers tiled along the sequence
}
\examples{
km <- get_kmers(sequence = "MRSLLFVVGAWVAALVTNLTPDAALASGTTTTAAAGNTSATASPGDNATSIDAGST",
                kmer_length = 12,
                overlap = 3)
print(km)

}
\seealso{
\code{\link{make_kmer_dataframe}}
}
\author{
Jennifer L. Remmel, Siyang Xia \email{sxia@hsph.harvard.edu}
}
\keyword{manipulation}
\keyword{peptide}
\keyword{sequence}
