% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kmer_dataframe.R
\name{make_kmer_dataframe}
\alias{make_kmer_dataframe}
\title{Generate a kmer data frame}
\usage{
make_kmer_dataframe(peptides, kmer_length = 12, overlap = 6)
}
\arguments{
\item{peptides}{a tibble of peptides with two columns named "id" and "pep_aa"}

\item{kmer_length}{length of desired kmer}

\item{overlap}{number of overlapped residuals between two consecutive kmers}
}
\value{
a data frame of kmers with the desired structure for alignment
}
\description{
convert a peptide data frame to kmer data frame
}
\examples{
data(peptide_df)
kmer_df <- make_kmer_dataframe(peptides = peptide_df)
head(kmer_df)

}
\seealso{
\code{\link{get_kmers}}
}
\author{
Jennifer L. Remmel, Siyang Xia \email{sxia@hsph.harvard.edu}
}
\keyword{manipulation}
\keyword{peptide}
\keyword{sequence}
