% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_comparisons_elmt.R
\name{peptide_comparisons_elmt}
\alias{peptide_comparisons_elmt}
\title{Peptide pairwise alignment - element-wise version}
\usage{
peptide_comparisons_elmt(
  peptides,
  sub_matrix = "BLOSUM50",
  gap_opening = 10,
  gap_extension = 4,
  align_type = "local",
  self_comparison = TRUE,
  full_align = FALSE,
  other_info = TRUE,
  parallel_ncore = 1
)
}
\arguments{
\item{peptides}{a tibble of peptides with two columns named "id" and "pep_aa"}

\item{sub_matrix}{substitution matrix to use for the alignment, default = "BLOSUM50"}

\item{gap_opening}{penalty for starting a gap in the pairwiseAlignment function, default = 10}

\item{gap_extension}{penalty for extending a gap in the pairwiseAlignment function, default = 4}

\item{align_type}{alignment type in the pairwiseAlignment function,
options including "local" (default), "global", "overlap", "global-local", and "local-global".}

\item{self_comparison}{whether alignment with self is performed, default = TRUE}

\item{full_align}{whether to keep the full alignment results as a column in the final results, default = FALSE}

\item{other_info}{whether to keep peptide information in the output, default = TRUE}

\item{parallel_ncore}{number of cores used for internal parallel in data.table}
}
\value{
a tibble with sequence alignment results from all unique pairs of peptides
}
\description{
perform pairwise sequence alignment for a dataset of peptides using dplyr and vectorized
pairwiseAlignment function in the package "Biostrings"
}
\examples{
data(peptide_df)
pairwise_align1 <- peptide_comparisons_elmt(peptides = peptide_df)
pairwise_align2 <- peptide_comparisons_elmt(peptides = peptide_df, parallel_ncore = 4)
head(pairwise_align1)
head(pairwise_align2)

}
\references{
pairwiseAlignment: \url{https://bioconductor.org/packages/devel/bioc/vignettes/Biostrings/inst/doc/PairwiseAlignments.pdf}
}
\seealso{
\code{\link{peptide_comparisons_vctz}}
}
\author{
Jennifer L. Remmel, Siyang Xia \email{sxia@hsph.harvard.edu}
}
\keyword{alignment}
\keyword{peptide}
