% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_df.R
\docType{data}
\name{peptide_df}
\alias{peptide_df}
\title{Example data from VirScan peptide library}
\format{
An object of class \code{"tibble"}.
}
\source{
VirScan library
}
\usage{
data(peptide_df)
}
\description{
The dataset contains 20 peptides randomly sampled from the VirScan peptide library
It includes four columns: id, pep_aa, organism and org_group
}
\examples{
data(peptide_df)
head(peptide_df)
}
\keyword{datasets}
