% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_pairwise_cooccurrence_mc.R
\name{peptide_pairwise_cooccurrence_mc}
\alias{peptide_pairwise_cooccurrence_mc}
\title{Peptide pairwise co-occurrence}
\usage{
peptide_pairwise_cooccurrence_mc(
  d,
  test_method = "fisher",
  correlation = TRUE,
  p_adjust_method = "holm",
  parallel_ncore = 1
)
}
\arguments{
\item{d}{a data frame or tibble with columns indicating peptides and rows indicating samples.
column names are peptide id and row names are sample id}

\item{test_method}{method to perform statistical test, options: "fisher" or "chisq" or "none"}

\item{correlation}{whether phi correlation is calculated}

\item{p_adjust_method}{method to adjust multiple test p values
options: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{parallel_ncore}{number of cores used for internal parallel in data.table}
}
\value{
a tibble contains co-occurence analysis results of all pairs of peptides
}
\description{
perform co-occurrence analysis across all pairs of peptides
}
\examples{
data(peptide_z)
peptide_bi <- as.data.frame((peptide_z > 3.5) * 1)
pairwise_cooccur <- peptide_pairwise_cooccurrence(d = peptide_bi)
head(pairwise_cooccur)

}
\seealso{
\code{\link{peptide_coocurrence}}
}
\author{
Siyang Xia \email{sxia@hsph.harvard.edu}
}
\keyword{co-occurrence}
\keyword{peptide}
