% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_z.R
\docType{data}
\name{peptide_z}
\alias{peptide_z}
\title{An example dataset of z-scores from VirScan analysis}
\format{
An object of class \code{"data.frame"}.
}
\source{
VirScan library
}
\usage{
data(peptide_z)
}
\description{
The dataset contains 50 peptides (column) and 12 samples (row) from the VirScan peptide library
}
\examples{
data(peptide_z)
head(peptide_z)
}
\keyword{datasets}
