% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pairwise.R
\name{to_pairwise}
\alias{to_pairwise}
\title{Enable pairwise analysis with customized function}
\usage{
to_pairwise(d, f, unique_pair = TRUE, same_comparison = FALSE, mc = 1)
}
\arguments{
\item{d}{a data frame or tibble or data table}

\item{f}{a function to perform some analysis. The inputs of the function should be two numeric vectors.}

\item{unique_pair}{a binary indicator of whether only unique pairs should be kept.
If \code{TRUE}, "B-A" pair will be removed if "A-B" pair is present.}

\item{same_comparison}{a binary indicator of whether to keep self pairs (i.e., "A-A").
If \code{TRUE}, self pairs will be kept.}

\item{mc}{number of cores to use for multi-thread parallel analysis. Pass to
\code{\link[parallel]{mclapply}}.
Deafult: 1}
}
\value{
a data table contains analysis results of all pairs of columns.
    The first two columns ("id1" and "id2") contain the two column names in \code{d}
    that were included in each pairwise analysis.
}
\description{
given a data frame, perform pairwise analysis with any function on all pairs of columns
}
\examples{
data(peptide_z)
peptide_bi <- as.data.frame((peptide_z > 5) * 1)
output <- to_pairwise(d = peptide_bi, f = peptide_cooccurrence)
head(output)

}
\author{
Siyang Xia \email{sxia@hsph.harvard.edu}
}
\keyword{analysis}
\keyword{column}
\keyword{pairwise}
