% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_cooccurrence.R
\name{peptide_cooccurrence}
\alias{peptide_cooccurrence}
\title{Summarize the co-occurrence of two peptides}
\usage{
peptide_cooccurrence(x1, x2, test_method = "fisher", correlation = TRUE)
}
\arguments{
\item{x1}{a logical vector of peptide 1's presence in multiple samples}

\item{x2}{a logical vector of peptide 2's presence in multiple samples}

\item{test_method}{method to perform statistical test, options: "fisher" or "chisq" or "none"}

\item{correlation}{whether phi correlation is calculated}
}
\value{
a vector that contains 4-7 elements (depending on whether to perform test and correlation):
  TT: frequency of both peptide positive
  TF: frequency of peptide 1 positive and peptide 2 negative
  FT: frequency of peptide 1 genative and peptide 2 position
  FF: frequency of both peptide negative
  stat: key statistics from the selected test (odds ratio for Fisher's exact test and chi-squared for chi-squared test)
  p: p value from the selected test
  phi: phi correlation coefficient (calculated by Pearson correlation)
}
\description{
calculate a contingency table of two peptides' presence in multiple individuals and
perform independency test and phi correlation coefficient
}
\examples{
data(peptide_z)
peptide_bi <- as.data.frame((peptide_z > 3.5) * 1)
cooccur <- peptide_cooccurrence(x1 = peptide_bi[, 1], x2 = peptide_bi[, 2])
print(cooccur)

}
\seealso{
\code{\link{peptide_pairwise_coocurrence}}
}
\author{
Siyang Xia \email{sxia@hsph.harvard.edu}
}
\keyword{co-occurrence}
\keyword{peptide}
