% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_pairwise_alignment.R
\name{peptide_pairwise_alignment}
\alias{peptide_pairwise_alignment}
\title{Peptide pairwise alignment - vectorized version}
\usage{
peptide_pairwise_alignment(
  peptides,
  sub_matrix = "BLOSUM50",
  gap_opening = 10,
  gap_extension = 4,
  align_type = "local",
  self_comparison = TRUE,
  full_align = FALSE,
  other_info = TRUE,
  parallel_ncore = NULL,
  output_str = "data.table"
)
}
\arguments{
\item{peptides}{a tibble of peptides with two columns named "id" and "pep_aa"}

\item{sub_matrix}{substitution matrix to use for the alignment, default = "BLOSUM50"}

\item{gap_opening}{penalty for starting a gap in the pairwiseAlignment function, default = 10}

\item{gap_extension}{penalty for extending a gap in the pairwiseAlignment function, default = 4}

\item{align_type}{alignment type in the pairwiseAlignment function,
options including "global", "local", "overlap", "global-local", and "local-global".}

\item{self_comparison}{whether alignment with self is performed, default = TRUE}

\item{full_align}{whether to keep the full alignment results as a column in the final results, default = FALSE}

\item{other_info}{whether to keep peptide information in the output, default = TRUE}

\item{parallel_ncore}{number of cores used for internal parallel in data.table}

\item{output_str}{format of the output to return, options: "data.table", "data.frame", and "tibble"}
}
\value{
a data frame/data table/tibble with sequence alignment results from all unique pairs of peptides
}
\description{
perform pairwise sequence alignment for a dataset of peptides taking advantange of data.table and vectorized
pairwiseAlignment function in the package "Biostrings"
}
\examples{
data(peptide_df)
pairwise_align <- peptide_pairwise_alignment(peptides = peptide_df)
head(pairwise_align)

}
\references{
pairwiseAlignment: \url{https://bioconductor.org/packages/devel/bioc/vignettes/Biostrings/inst/doc/PairwiseAlignments.pdf}
}
\seealso{
\code{\link[Biostrings]{pairwiseAlignment}}
}
\author{
Jennifer L. Remmel, Siyang Xia \email{sxia@hsph.harvard.edu}
}
\keyword{alignment}
\keyword{peptide}
